[CGD-94] Added data table to manage lights
There is a Light area that can be covered over a region.
In the data table you create a row, and select the light object and area it is in. This uses soft object references so it is a selection of the current objects in the scene.
LightArea checks through table for lights linked to the current area. Then activates/deactivates based on event.