[CGD-98] Loading levels!
functionality setup using one event in the game mode  to load an event! this ties with the results screen where it can replay and play the next level and will be compatible when we make the level select. just needs the actual load function but we have no levels to load so thats blocked