[CGD-106] on screen timer added and functional
added a timer on the screen for each level that will display ongoing time. atm there is a bug in the logic where it has a second at the end of each minute wrong? looking to add completion time and fix this next. if time ill add a very simple star system too :)